Not forcefully upgrading requirements.
If --no-cache and --upgrade is set then the pipeline will try to
re-install wxPython that is already present in the docker image used for
the build. This will in turn fail the build. Therefore we just:
* specify the requirements normally
* do not a force refresh install of the project's requirements
* and use the already installed version of wxPython

This is the same underlying reason for which we do not use a python
virtual environment in the build.